v1.12.0 process_eoi postcode restriction; implemented fields sorting for manage.php

+++ project2/manage_delete.php
@@ -12,24 +12,27 @@
 <?php
 	require_once 'settings.php';
 	session_start();
-	if (!empty($_SESSION['check_sql'])){
-	$check_sql = $_SESSION['check_sql'];
-	} else $check_sql='0';
-	if (!empty($_SESSION['searchq'])){
-	$search = $_SESSION['searchq'];
-	} else $search='0';
+	if (!isset($_SESSION['user_id'])) {
+		header("Location: login.php");
+		exit();
+	}	
+	if ($_SERVER["REQUEST_METHOD"] !== "POST") {
+		mysqli_close($conn);
+		header("Location: manage.php");
+		exit();
+	}
 	// mysqli_use_result($conn);
     if(isset($_POST['delete_button'])) {
         $search_result = mysqli_real_escape_string($conn, $search);
-        $sql1 = "DELETE FROM eoi WHERE
-									(reference_number LIKE '%$search_result%' or
-									first_name LIKE '%$search_result%' or
-									last_name LIKE '%$search_result%'
-									) and
-									($check_sql)
-        ";
-        $result = mysqli_query($conn, $sql1);
-		unset($_SESSION['searchq'], $_SESSION['check_sql'], $_SESSION['search_sql']);
+		foreach($_SESSION["delete_ids"] as $delete_id){
+			$sql7 = "DELETE FROM eoi_skills WHERE eoi_number = $delete_id";
+			mysqli_query($conn, $sql7);	
+			$sql1 = "DELETE FROM eoi WHERE
+										eoi_number = $delete_id
+			";
+        	mysqli_query($conn, $sql1);
+		}
+		unset($_SESSION['search_sql'], $_SESSION["delete_ids"]);
 		mysqli_close($conn);
         echo header('Location:manage.php');
     } else {
